+++ UI/Events.ts
@@ -9,7 +9,8 @@
 import * as Subscriber from './_events/Subscriber';
 import * as EventUtils from './_events/EventUtils';
 import { FastTouchEndController } from './_events/Touch/FastTouchEndController';
-import { WasabyEvents } from './_events/WasabyEvents';
+import  WasabyEventsReact from './_events/WasabyEventsReact';
+import  WasabyEventsInferno from './_events/WasabyEventsInferno';
 import { callNotify } from './_events/Notify';
 
 export { ISyntheticEvent, IWasabyEventSystem, IWasabyEvent } from './_events/IEvents';
@@ -20,6 +21,7 @@ export {
     Subscriber,
     EventUtils,
     FastTouchEndController,
-    WasabyEvents,
+    WasabyEventsReact,
+    WasabyEventsInferno,
     callNotify
 };
