+++ UI/_events/IEvents.ts
@@ -1,4 +1,4 @@
-import { IControlNode, IDOMEnvironment, IWasabyHTMLElement, TModifyHTMLNode } from 'UI/_vdom/Synchronizer/interfaces';
+import { IControlNode, IDOMEnvironment, IWasabyHTMLElement } from 'UI/_vdom/Synchronizer/interfaces';
 import {Control} from '../_react/Control/WasabyOverReact';
 
 type VoidFunction = () => void;
@@ -85,8 +85,6 @@ export interface IWasabyEvent {
 }
 
 export interface IWasabyEventSystem {
-    useWasabyOverReact: boolean;
-    initWasabyEventSystem: (rootNode: TModifyHTMLNode,  environment: IDOMEnvironment, tabKeyHandler?: Function) => void;
     captureEventHandler: (event: Event) => void;
     callEventsToDOM: VoidFunction;
     handleSpecialEvent: (eventName: string, eventHandler: Function, environment: IDOMEnvironment) => void;
