+++ UI/_react/Control/interfaces.ts
@@ -1,5 +1,4 @@
 import { _IGeneratorType } from 'UI/Executor';
-import { IWasabyEventSystem } from 'UI/Events';
 
 export type TemplateFunction = (
    data: object,
@@ -14,7 +13,6 @@ export type TemplateFunction = (
 export interface IControlOptions {
    readOnly?: boolean;
    theme?: string;
-   eventSystem?: IWasabyEventSystem;
 }
 
 export interface IControlState {
